# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 伴AI API
  version: 1.27.0
  description: M1-M4 platform plus the M6 reliability control plane, governed Agent Studio, unified adjustable usage ledger, multi-instance runtime-configuration convergence, persisted cost and quality analytics with time trends, predictive budget burn and guarded model-switch recommendations, per-budget proactive forecast policies with sample gating, audited human recommendation feedback, revision-bound application and rollback-execution tracking, configurable before/after observation windows, downloadable periodic budget reviews, read-only rollback recommendations and post-rollback verification with acknowledged/closed human disposition, audited budget guardrails, automatic budget incidents with escalation plus email or management-console recovery notifications, baseline anomaly detection, Loki-backed redacted structured system logs with PostgreSQL fallback and Run/Trace correlation, threshold alert evaluation, downloadable incident evidence, durable W3C/OpenTelemetry Trace Context propagation across HTTP, Outbox, Kafka, Go/Python Agent workers and Langfuse, optional fail-open Langfuse LLM/Agent traces with generations, node observations and scores, operator DLQ replay, and compensation records. Every HTTP response includes canonical traceparent and X-Trace-ID correlation headers; clients may send traceparent/tracestate, while X-Trace-ID remains a normalized compatibility input.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/agent-runs/{run_id}:
    get:
      operationId: getAgentRun
      tags: [Agent]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: run_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: User-owned Agent Run state. Confirmation tokens and internal input are never exposed.
          content: { application/json: { schema: { $ref: "#/components/schemas/AgentRun" } } }
        "404": { description: Run does not exist or belongs to another user. }
...
  /v1/ops/agent-runs:
    get:
      operationId: listOperatorAgentRuns
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Lists privacy-reduced Agent Run metadata. Prompts, responses, tool arguments, and tool results are not returned.
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [accepted, queued, running, waiting_approval, waiting_tool, completed, failed, cancel_requested, cancelled, timed_out] } }
        - { name: module, in: query, required: false, schema: { type: string, enum: [companion, life, work] } }
        - { name: error_code, in: query, required: false, schema: { type: string, maxLength: 128 } }
        - { name: q, in: query, required: false, description: Exact run, thread, or conversation identifier., schema: { type: string, maxLength: 128 } }
        - { name: from, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: false, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        "200":
          description: Agent Run summaries ordered newest first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, next_cursor]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/AgentRunOperationsSummary" } }
                  next_cursor: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /v1/ops/agent-runs/{run_id}:
    get:
      operationId: getOperatorAgentRun
      tags: [Operations]
      security: [{ operatorBearerAuth: [] }]
      description: Returns a redacted trace. User prompts, assistant responses, tool arguments/results, and confirmation secrets are excluded.
      parameters:
        - { name: run_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Operational run detail with node trace, model usage, tool metadata, and state events.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentRunOperationsDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    operatorBearerAuth:
      type: http
      scheme: bearer
      description: Uses either legacy OPERATOR_TOKEN or an operator account token. When operator MFA is enabled, clients must also send X-Operator-TOTP.
  responses:
    Unauthorized:
      description: Authentication is missing, invalid, expired, or revoked.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationError:
      description: Request failed domain validation.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource was not found in the caller scope.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    AgentRun:
      type: object
      required: [id, conversation_id, character_id, module, status, revision, deadline_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        character_id: { type: string, format: uuid }
        module: { type: string, enum: [companion, life, work] }
        agent_definition_key: { type: string, description: Published Agent Studio definition captured when the run was accepted. }
        agent_definition_version_id: { type: string, format: uuid }
        agent_definition_version: { type: integer, minimum: 1 }
        agent_definition_revision: { type: integer, minimum: 1 }
        agent_definition_fingerprint: { type: string, pattern: '^[0-9a-f]{64}$' }
        model_profile_version_id: { type: string, format: uuid, description: Immutable model profile version captured when the run was accepted. }
        model_profile_config_version: { type: string }
        model_profile_fingerprint: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { type: string, enum: [accepted, queued, running, waiting_approval, waiting_tool, completed, failed, cancel_requested, cancelled, timed_out] }
        revision: { type: integer, minimum: 1 }
        deadline_at: { type: string, format: date-time, description: "Fixed total deadline; retries and approval resumes do not extend it." }
        output:
          type: object
          additionalProperties: true
          properties:
            response: { type: string }
            interrupts:
              type: array
              maxItems: 1
              items:
                type: object
                required: [type, summary]
                properties:
                  type: { type: string, enum: [tool_approval] }
                  summary: { type: string }
                  tool_name: { type: string }
                  risk_level: { type: string, enum: [none, low, medium, high] }
                  arguments: { type: object, additionalProperties: true }
        error_code: { type: string }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    AgentRunOperationsSummary:
      type: object
      additionalProperties: false
      required: [id, thread_id, user_id, conversation_id, module, graph_name, graph_version, status, revision, model_calls, prompt_tokens, completion_tokens, cost_micros, tool_calls, duration_ms, deadline_at, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        thread_id: { type: string }
        user_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        module: { type: string, enum: [companion, life, work] }
        graph_name: { type: string }
        graph_version: { type: string }
        agent_definition_key: { type: string }
        agent_definition_version_id: { type: string, format: uuid }
        agent_definition_version: { type: integer, minimum: 1 }
        agent_definition_revision: { type: integer, minimum: 1 }
        agent_definition_fingerprint: { type: string, pattern: '^[0-9a-f]{64}$' }
        model_profile_key: { type: string }
        model_profile_version_id: { type: string, format: uuid }
        model_profile_revision: { type: integer, minimum: 1 }
        model_profile_config_version: { type: string }
        model_profile_fingerprint: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { type: string, enum: [accepted, queued, running, waiting_approval, waiting_tool, completed, failed, cancel_requested, cancelled, timed_out] }
        current_node: { type: string }
        revision: { type: integer, minimum: 1 }
        model_calls: { type: integer, minimum: 0 }
        prompt_tokens: { type: integer, minimum: 0 }
        completion_tokens: { type: integer, minimum: 0 }
        cost_micros: { type: integer, minimum: 0 }
        tool_calls: { type: integer, minimum: 0 }
        error_code: { type: string }
        error_message: { type: string, description: Truncated operational error text. }
        duration_ms: { type: integer, minimum: 0 }
        deadline_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
...
    AgentRunOperationsModelCall:
      type: object
      additionalProperties: true
      description: Governed model-call metadata without prompt or completion content.
      properties:
        graph_node: { type: string }
        role: { type: string }
        status: { type: string }
        provider: { type: string }
        requested_model: { type: string }
        returned_model: { type: string }
        prompt_tokens: { type: integer, minimum: 0 }
        completion_tokens: { type: integer, minimum: 0 }
        cached_tokens: { type: integer, minimum: 0 }
        reasoning_tokens: { type: integer, minimum: 0 }
        cost_micros: { type: integer, minimum: 0 }
        latency_ms: { type: integer, minimum: 0 }
        timeout_ms: { type: integer, minimum: 0 }
        error_status: { type: string }
        retryable: { type: boolean }
        contract_valid: { type: boolean }
        contract_error: { type: string }
    AgentRunOperationsNode:
      type: object
      additionalProperties: false
      properties:
        node: { type: string }
        status: { type: string }
        duration_ms: { type: integer, minimum: 0 }
        details: { type: object, additionalProperties: true, description: Recursively redacted bounded diagnostic metadata. }
    AgentRunOperationsToolCall:
      type: object
      additionalProperties: false
      required: [id, run_id, tool_name, risk_level, status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        tool_name: { type: string }
        risk_level: { type: string, enum: [none, low, medium, high] }
        status: { type: string, enum: [proposed, approved, running, succeeded, failed, cancelled] }
        error_code: { type: string }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
    AgentRunOperationsEvent:
      type: object
      additionalProperties: false
      required: [id, sequence, type, created_at]
      properties:
        id: { type: integer, format: int64 }
        sequence: { type: integer, format: int64 }
        type: { type: string }
        created_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: true, description: Allowlisted state-transition metadata. }
    AgentRunOperationsDetail:
      type: object
      additionalProperties: false
      required: [run, graph, budget, model_manifest, model_calls, node_trace, langfuse, tool_calls, events]
      properties:
        run: { $ref: "#/components/schemas/AgentRunOperationsSummary" }
        graph: { type: object, additionalProperties: true }
        budget: { type: object, additionalProperties: true }
        model_manifest: { type: object, additionalProperties: true }
        model_calls: { type: array, items: { $ref: "#/components/schemas/AgentRunOperationsModelCall" } }
        node_trace: { type: array, items: { $ref: "#/components/schemas/AgentRunOperationsNode" } }
        langfuse: { $ref: "#/components/schemas/LangfuseRunReference" }
        tool_calls: { type: array, items: { $ref: "#/components/schemas/AgentRunOperationsToolCall" } }
        events: { type: array, items: { $ref: "#/components/schemas/AgentRunOperationsEvent" } }
...
    LangfuseRunReference:
      type: object
      additionalProperties: false
      required: [enabled]
      properties:
        enabled: { type: boolean }
        trace_id: { type: string, pattern: "^[0-9a-f]{32}$" }
        initialized: { type: boolean }
        capture_content: { type: boolean }
        environment: { type: string }
...